Move common config utils to helper
